Deploy openssl config with ciphers and ChaCha20
The /etc/ssl/openssl.cnf config file, which is used for the config
of the openssl library, wich in turn is used by webservers, allows
to specify advanced options, that cannot be configured in apache2
or nginx (yet). Using the following options

Options = ServerPreference,PrioritizeChaCha

has the effect, that the server preferred cipher order is forced,
except for clients that prefer CHACHA20_POLY1305_SHA256. This is
the case for mobile devices that do not have AES-NI CPU support.
ChaCha20 is roughly 3 times faster than AES-128 on such devices.
The difference is more dramatic on older and less powerful phones.