explicit column_case_sensitive for mysql_user module
Fixes ansible warning: "Option column_case_sensitive is not provided.
The default is now false, so the column's name will be uppercased. The
default will be changed to true in community.mysql 4.0.0".
We track the current "False" so that the module update won't change
anything. We may later switch to the new default "True" on-demand.